Helper functions for authentification refactoring
Move helper functions for authentification to
separate file. Add unit test for time parsing.

Relates-To: OLPEDGE-1921
