+++ olp-cpp-sdk-authentication/tests/CMakeLists.txt
@@ -17,6 +17,7 @@
 
 set(OLP_AUTHENTICATION_TEST_SOURCES
     AuthenticationCredentialsTest.cpp
+    AuthenticationClientTest.cpp
     DecisionApiClientTest.cpp
 )
